+++ src/main/scala/ai/privado/languageEngine/java/tagger/PrivadoTagger.scala
@@ -40,8 +40,8 @@ import ai.privado.languageEngine.java.tagger.collection.{
   SOAPCollectionTagger
 }
 import ai.privado.languageEngine.java.tagger.config.JavaDBConfigTagger
-import ai.privado.languageEngine.java.tagger.sink.api.JavaAPISinkTagger
-import ai.privado.languageEngine.java.tagger.sink.{InheritMethodTagger, JavaAPITagger, MessagingConsumerCustomTagger}
+import ai.privado.languageEngine.java.tagger.sink.api.{JavaAPISinkTagger, JavaAPITagger}
+import ai.privado.languageEngine.java.tagger.sink.{InheritMethodTagger, MessagingConsumerCustomTagger}
 import ai.privado.languageEngine.java.tagger.source.*
 import ai.privado.tagger.PrivadoBaseTagger
 import ai.privado.tagger.collection.{AndroidCollectionTagger, WebFormsCollectionTagger}
@@ -83,9 +83,8 @@ class PrivadoTagger(cpg: Cpg) extends PrivadoBaseTagger {
 
     new JavaS3Tagger(cpg, s3DatabaseDetailsCache).createAndApply()
 
-    JavaAPISinkTagger.applyTagger(cpg, ruleCache, privadoInputConfig)
+    JavaAPISinkTagger.applyTagger(cpg, ruleCache, privadoInputConfig, appCache)
 
-    new JavaAPITagger(cpg, ruleCache, privadoInputConfig, appCache).createAndApply()
     // Custom Rule tagging
     if (!privadoInputConfig.ignoreInternalRules) {
       // Adding custom rule to cache
