+++ src/main/scala/ai/privado/languageEngine/java/tagger/sink/api/JavaAPISinkTagger.scala
@@ -1,8 +1,7 @@
 package ai.privado.languageEngine.java.tagger.sink.api
 
-import ai.privado.cache.RuleCache
+import ai.privado.cache.{AppCache, RuleCache}
 import ai.privado.entrypoint.PrivadoInput
-import ai.privado.languageEngine.java.tagger.sink.JavaAPIRetrofitTagger
 import ai.privado.tagger.sink.api.APISinkTagger
 import io.shiftleft.codepropertygraph.generated.Cpg
 
@@ -12,9 +11,9 @@ object JavaAPISinkTagger extends APISinkTagger {
     * @param cpg
     * @param ruleCache
     */
-  override def applyTagger(cpg: Cpg, ruleCache: RuleCache, privadoInput: PrivadoInput): Unit = {
+  override def applyTagger(cpg: Cpg, ruleCache: RuleCache, privadoInput: PrivadoInput, appCache: AppCache): Unit = {
 
-    super.applyTagger(cpg, ruleCache, privadoInput)
+    super.applyTagger(cpg, ruleCache, privadoInput, appCache)
     new JavaAPIRetrofitTagger(cpg, ruleCache).createAndApply()
 
     if (privadoInput.enableAPIByParameter) {
@@ -23,6 +22,8 @@ object JavaAPISinkTagger extends APISinkTagger {
     }
 
     new JavaAPISinkByMethodFullNameTagger(cpg, ruleCache).createAndApply()
+
+    new JavaAPITagger(cpg, ruleCache, privadoInput, appCache).createAndApply()
   }
 
 }
