+++ src/main/scala/ai/privado/languageEngine/kotlin/tagger/PrivadoTagger.scala
@@ -6,8 +6,8 @@ import ai.privado.feeder.PermissionSourceRule
 import ai.privado.languageEngine.java.feeder.StorageInheritRule
 import ai.privado.languageEngine.java.tagger.collection.{CollectionTagger, MethodFullNameCollectionTagger}
 import ai.privado.languageEngine.java.tagger.config.JavaDBConfigTagger
-import ai.privado.languageEngine.java.tagger.sink.api.JavaAPISinkTagger
-import ai.privado.languageEngine.java.tagger.sink.{InheritMethodTagger, JavaAPITagger}
+import ai.privado.languageEngine.java.tagger.sink.api.{JavaAPISinkTagger, JavaAPITagger}
+import ai.privado.languageEngine.java.tagger.sink.InheritMethodTagger
 import ai.privado.languageEngine.java.tagger.source.*
 import ai.privado.languageEngine.kotlin.feeder.StorageAnnotationRule
 import ai.privado.languageEngine.kotlin.tagger.sink.StorageAnnotationTagger
@@ -59,9 +59,7 @@ class PrivadoTagger(cpg: Cpg) extends PrivadoBaseTagger {
       new StorageAnnotationTagger(cpg, ruleCache).createAndApply()
     }
 
-    JavaAPISinkTagger.applyTagger(cpg, ruleCache, privadoInputConfig)
-
-    new JavaAPITagger(cpg, ruleCache, privadoInputConfig, appCache = appCache).createAndApply()
+    JavaAPISinkTagger.applyTagger(cpg, ruleCache, privadoInputConfig, appCache)
 
     new AndroidCollectionTagger(
       cpg,
