+++ src/main/scala/ai/privado/tagger/sink/api/APISinkTagger.scala
@@ -1,12 +1,12 @@
 package ai.privado.tagger.sink.api
 
-import ai.privado.cache.RuleCache
+import ai.privado.cache.{AppCache, RuleCache}
 import ai.privado.entrypoint.PrivadoInput
 import io.shiftleft.codepropertygraph.generated.Cpg
 
 trait APISinkTagger {
 
-  def applyTagger(cpg: Cpg, ruleCache: RuleCache, privadoInput: PrivadoInput): Unit = {
+  def applyTagger(cpg: Cpg, ruleCache: RuleCache, privadoInput: PrivadoInput, appCache: AppCache): Unit = {
     new InferenceAPIEndpointTagger(cpg, ruleCache).createAndApply()
   }
